+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/modinit/BzTags.java
@@ -196,6 +196,7 @@ public class BzTags {
     public static final TagKey<MobEffect> LIFE_CURE_EFFECTS = TagKey.create(Registries.MOB_EFFECT, new ResourceLocation(Bumblezone.MODID, "essence/life/cure_effects"));
     public static final TagKey<MobEffect> RADIANCE_SUN_EFFECTS = TagKey.create(Registries.MOB_EFFECT, new ResourceLocation(Bumblezone.MODID, "essence/radiance/sun_effects"));
     public static final TagKey<MobEffect> RAGING_RAGE_EFFECTS = TagKey.create(Registries.MOB_EFFECT, new ResourceLocation(Bumblezone.MODID, "essence/raging/rage_effects"));
+    public static final TagKey<MobEffect> FOG_ADJUSTING_EFFECTS = TagKey.create(Registries.MOB_EFFECT, new ResourceLocation(Bumblezone.MODID, "fog_adjusting_effects"));
 
     public static final TagKey<PoiType> IS_NEAR_BEEHIVE_ADVANCEMENT_TRIGGER_POI = TagKey.create(Registries.POINT_OF_INTEREST_TYPE, new ResourceLocation(Bumblezone.MODID, "is_near_beehive_advancement_trigger"));